File rename, remove comments
