+++ app/models/post_policy.rb
@@ -54,7 +54,7 @@ class PostPolicy < ActiveRecord::Base
   def add_users_group
     return if add_users_to_group.nil?
 
-    group = Group.find_by(id: add_users_to_group)
+    Group.find_by(id: add_users_to_group)
   end
 
   private
@@ -118,17 +118,18 @@ end
 #
 # Table name: post_policies
 #
-#  id               :bigint           not null, primary key
-#  post_id          :bigint           not null
-#  renew_start      :datetime
-#  renew_days       :integer
-#  next_renew_at    :datetime
-#  reminder         :string
-#  last_reminded_at :datetime
-#  version          :string
-#  created_at       :datetime         not null
-#  updated_at       :datetime         not null
-#  renew_interval   :integer
-#  private          :boolean          default(FALSE), not null
-#  last_bumped_at   :datetime
+#  id                 :bigint           not null, primary key
+#  post_id            :bigint           not null
+#  renew_start        :datetime
+#  renew_days         :integer
+#  next_renew_at      :datetime
+#  reminder           :string
+#  last_reminded_at   :datetime
+#  version            :string
+#  created_at         :datetime         not null
+#  updated_at         :datetime         not null
+#  renew_interval     :integer
+#  private            :boolean          default(FALSE), not null
+#  last_bumped_at     :datetime
 #  add_users_to_group :integer
+#
